requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
groq